% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acadsleep}
\alias{acadsleep}
\title{Sleep and Academic Success Dataset}
\format{
A data frame with 1338 rows and 10 variables:
\describe{
  \item{participantID}{unique participant identification.}
  \item{gender}{indicator of participant's self-identified gender.}
  \item{PSS}{score on the Perceived Stress Scale (0-40). Higher values indicate higher stress levels.}
  \item{PSQI}{score on the Pittsburgh Sleep Quality Index (0-21). Higher values indicate worse sleep quality.}
  \item{PHQ9}{score on the Patient Health Questionnaire (0-27) for assessing depression. Higher values indicate more severe depression.}
  \item{FS}{score on the Flourishing Scale (8-56) assessing psychological well-being. Higher values indicate healthier well-being.}
  \item{cumulGPA}{participant's cumulative GPA on a 4-point scale.}
  \item{springGPA}{participant's GPA (4-point scale) during only the semester for which the study was conducted.}
  \item{sleepHours}{the number of hours participant slept the previous night.}
  \item{sleepRate}{indicator rating of the sleep quality for the previous night.}
}
}
\source{
<https://osf.io/nrd4g>
}
\usage{
acadsleep
}
\description{
Data from a study examining the relationships between daily sleep ratings, well-being, and academic success over the course of a college semester.
}
\details{
There were several records where the PSQI was noted as 28; since the maximum value for the PSQI is 21, these values were set to missing in this dataset. There were a few individuals who recorded "0" for amount of sleep, but this was not an option in the survey; so, these were set to missing in this dataset.
}
\keyword{datasets}
