% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_compute_p.R
\name{boot_compute_p}
\alias{boot_compute_p}
\title{Compute p-value from a bootstrap sample of test statistics.}
\usage{
boot_compute_p(bootobj)
}
\arguments{
\item{bootobj}{vector containing the bootstrap test statistics returned by
\code{\link{boot_residual_p}}.}
}
\value{
data.frame containing an ANOVA table comparing the two models. The
p-value is determined from the bootstrap test statistics.
}
\description{
Given bootstrap estimates of a test statistic, compute a p-value.
}
\details{
The "two-sided" p-value is computed in that the test statistic considered is
the F-statistic from a nested F-test.
}
\examples{
\dontrun{
test.df <- data.frame(x = seq(10), y = seq(10) + rnorm(10))
fit1 <- lm(y ~ x, data = test.df)
fit0 <- lm(y ~ 1, data = test.df)
bootstrap_test_stats <- boot_residual_p(fit1, fit0, reps = 4999, wild = F)
boot_compute_p(bootstrap_test_stats)
}

}
\seealso{
\code{\link[stats]{anova}}
}
