% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_residual_ci.R
\name{boot_residual_ci}
\alias{boot_residual_ci}
\title{Perform (wild) residual bootstrap for linear model.}
\usage{
boot_residual_ci(fit, reps, wild, construct)
}
\arguments{
\item{fit}{\code{lm} object defining the model for which the parameters are
estimated.}

\item{reps}{scalar; number of bootstrap replications to perform.}

\item{wild}{boolean; if \code{TRUE}, a wild bootstrap is performed. If
\code{FALSE} a traditional residual bootstrap is performed.}

\item{construct}{string defining the type of construct to use when generating
from the distribution for the wild bootrap (see \code{\link{rmammen}}). If
\code{wild = FALSE}, this is ignored.}
}
\value{
matrix with the same number of rows as coefficients in \code{fit} and
\code{reps} columns. Each row contains the bootstrap estimates of the
corresponding parameters.
}
\description{
This is the workhorse for the \code{estimate_parameters()} function.
Given a linear model fit, performs a residual bootstrap (or wild bootstrap)
to estimate the sampling distribution of each of the parameters in the model.
This is not meant to be called by the user directly.
}
\examples{
\dontrun{
test.df <- data.frame(x = seq(10), y = seq(10) + rnorm(10))
fit <- lm(y ~ x, data = test.df)
boot_residual_ci(fit, reps = 4999, wild = FALSE)
}

}
