% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_compute_p.R
\name{bootstrap_compute_p}
\alias{bootstrap_compute_p}
\title{Compute p-value from a bootstrap sample of test statistics.}
\usage{
bootstrap_compute_p(bootobj)
}
\arguments{
\item{bootobj}{vector containing the bootstrap test statistics returned by a
workhorse bootstrap function under the null hypothesis:
\itemize{
  \item bootstrap_residual_null
  \item bootstrap_parametric_linear_null
  \item bootstrap_parametric_null
  \item bootstrap_case_null}
There should be one row for each parameter and each column contains a
bootstrap estimate.}
}
\value{
data.frame containing an ANOVA table comparing the two models. The
p-value is determined from the bootstrap test statistics.
}
\description{
Given bootstrap estimates of a test statistic, compute a p-value (always the
"two-sided" p-value).
}
\seealso{
\code{\link[stats]{anova}}
}
