% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_glm.R
\name{bootstrap_parametric}
\alias{bootstrap_parametric}
\title{Perform parametric bootstrap for generalized linear model.}
\usage{
bootstrap_parametric(fit, reps)
}
\arguments{
\item{fit}{\code{glm} object defining the model for which the parameters are
estimated.}

\item{reps}{scalar; number of bootstrap replications to perform.}
}
\value{
matrix with the same number of rows as coefficients in \code{fit} and
\code{reps} columns. Each row contains the bootstrap estimates of the
corresponding parameters.
}
\description{
This is the workhorse for the [estimate_parameters()] function when a
generalized linear model is specified and parametric bootstrap is required.
This is not meant to be called by the user directly.
}
\examples{
\dontrun{
fit <- glm(mpg ~ hp, data = mtcars)
bootstrap_parametric(fit, reps = 4999)
}

}
