% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lm_null.R
\name{bootstrap_parametric_linear_null}
\alias{bootstrap_parametric_linear_null}
\title{Compare nested linear models via parametric bootstrap.}
\usage{
bootstrap_parametric_linear_null(fit1, fit0, reps)
}
\arguments{
\item{fit1}{\code{lm} object defining the full model.}

\item{fit0}{\code{lm} object defining the reduced model.}

\item{reps}{scalar; number of bootstrap replications to perform.}
}
\value{
matrix with the same number of rows as coefficients in \code{fit} and
\code{reps} columns. Each row contains the bootstrap estimates of the
corresponding parameters.
}
\description{
This is the workhorse for the \code{compare_models()} function when a linear
model is specified without constant variance. Given two linear models, one
nested within the other, the null distribution of the test statistic
comparing the two models is estimated via a parametric bootstrap.
This is not meant to be called by the user directly.
}
\examples{
\dontrun{
fit1 <- lm(mpg ~ hp, data = mtcars)
fit0 <- lm(mpg ~ 1, data =  mtcars)

boot_parametric_linear_null(fit1, fit0, reps = 4999)
}

}
