% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_glm_null.R
\name{bootstrap_parametric_null}
\alias{bootstrap_parametric_null}
\title{Compare nested generalized linear models via parametric bootstrapping.}
\usage{
bootstrap_parametric_null(fit1, fit0, reps)
}
\arguments{
\item{fit1}{\code{glm} object defining the full model.}

\item{fit0}{\code{glm} object defining the reduced model.}

\item{reps}{scalar; number of bootstrap replications to perform.}
}
\value{
vector of length \code{reps} containing the test statistic from each
bootstrap replication. It also has an attribute containing an ANOVA table
comparing the two models.
}
\description{
This is the workhorse for the [compare_models()] function when a
generalized linear model is specified and parametric bootstrap is required.
This is not meant to be called by the user directly.
}
\examples{
\dontrun{
fit1 <- glm(mpg ~ hp, data = mtcars)
fit0 <- glm(mpg ~ 1, data = mtcars)

bootstrap_parametric_null(fit, fit0, reps = 4999)
}

}
