% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lm.R
\name{bootstrap_residual}
\alias{bootstrap_residual}
\title{Perform (optionally wild) residual bootstrap for linear model.}
\usage{
bootstrap_residual(fit, reps, wild, construct)
}
\arguments{
\item{fit}{\code{lm} object defining the model for which the parameters are
estimated.}

\item{reps}{scalar; number of bootstrap replications to perform.}

\item{wild}{boolean; if \code{TRUE}, a wild bootstrap is performed. If
\code{FALSE} a traditional residual bootstrap is performed.}

\item{construct}{string defining the type of construct to use when generating
from the distribution for the wild bootstrap (see [rmammen()]). If
\code{wild = FALSE}, this is ignored.}
}
\value{
matrix with the same number of rows as coefficients in \code{fit} and
\code{reps} columns. Each row contains the bootstrap estimates of the
corresponding parameters.
}
\description{
This is the workhorse for the [estimate_parameters()] function when a
linear model is specified. Given a linear model fit, performs a residual
bootstrap (or optionally wild bootstrap) to estimate the standard error of
and compute confidence intervals for each of the parameters in the model.
This is not meant to be called by the user directly.
}
