% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lm_null.R
\name{bootstrap_residual_null}
\alias{bootstrap_residual_null}
\title{Compare nested linear models via (optionally wild) residual bootstrap.}
\usage{
bootstrap_residual_null(fit1, fit0, reps, wild, construct)
}
\arguments{
\item{fit1}{\code{lm} object defining the full model.}

\item{fit0}{\code{lm} object defining the reduced model.}

\item{reps}{scalar; number of bootstrap replications to perform.}

\item{wild}{boolean; if \code{TRUE}, a wild bootstrap is performed. If
\code{FALSE} a traditional residual bootstrap is performed.}

\item{construct}{string defining the type of construct to use when generating
from the distribution for the wild bootstrap (see \code{\link{rmammen}}). If
\code{wild = FALSE}, this is ignored.}
}
\value{
vector of length \code{reps} containing the test statistic from each
bootstrap replication. It also has an attribute containing an ANOVA table
comparing the two models.
}
\description{
This is the workhorse for the \code{compare_models()} function when a linear
model is specified. Given two linear models, one nested within the other, the
null distribution of the test statistic comparing the two models is estimated
via a residual bootstrap (or wild bootstrap). This is not meant to be called
by the user directly.
}
\examples{
\dontrun{
fit1 <- lm(mpg ~ hp, data = mtcars)
fit0 <- lm(mpg ~ 1, data =  mtcars)

boot_residual_null(fit1, fit0, reps = 4999, wild = FALSE)
}

}
