% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellphones}
\alias{cellphones}
\title{Cellphone Addiction}
\format{
A data frame with 144 rows and 7 variables:
\describe{
  \item{Participant}{unique participant identification.}
  \item{Condition}{treatment condition assigned. "On" indicates access to cell phone, but asked to put it away; "Off" indicates access to cell phone, but asked to turn it off and put it away; "Removed" indicates cell phone was shut off and removed from participant's possession.}
  \item{PANAS1pos}{score on the Positive And Negative Affect Schedule indicating positive affect (10-50) prior to intervention. Higher scores indicate more positive emotions.}
  \item{PANAS2pos}{score on the Positive And Negative Affect Schedule indicating positive affect (10-50) following intervention. Higher scores indicate more positive emotions.}
  \item{PANAS1neg}{score on the Positive And Negative Affect Schedule indicating negative affect (10-50) prior to intervention. Higher scores indicate more negative emotions.}
  \item{PANAS2neg}{score on the Positive And Negative Affect Schedule indicating negative affect (10-50) following intervention. Higher scores indicate more negative emotions.}
  \item{Duration}{participant's perception of how long (minutes) they were left alone in a room following intervention (0-10). Actual time left alone was 3 minutes.}
}
}
\source{
<https://osf.io/xfkh8>
}
\usage{
cellphones
}
\description{
Data from a study examining the effect of being denied access to a cell phone.
}
\keyword{datasets}
