% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consent}
\alias{consent}
\title{Sexual Consent Dataset}
\format{
A data frame with 164 rows and 11 variables:
\describe{
  \item{SubjID}{unique subject identifier.}
  \item{Gender}{indicator of participant's self-identified gender. While several categories were initially offered on the survey, the only options selected were male/female.}
  \item{Age}{age of participant, years.}
  \item{Ethnicity}{indicator of participant's self-identified race/ethnicity.}
  \item{SexualOrientation}{indicator of participant's self-identified sexual orientation.}
  \item{RelationshipStatus}{indicator of participant's current relationship status.}
  \item{SexPartners}{number of previous sexual partners.}
  \item{ATSS}{total sexual attitudes score, with higher values indicating more liberal views on sexuality. Scores can range from 13 to 65.}
  \item{SexEd}{indicator of the coverage of "consent" in participant's high school sexual education.}
  \item{IndirectScore}{average of responses to 6-item Indirect Behavioral Approach to Consent subscale of the SCS-R. Values can range from 1 to 7, with higher values indicating appropriate consent practices.}
  \item{Consent}{average of the percent of time the participant has given and received verbal/nonverbal consent in past sexual encounters.}
}
}
\source{
<https://osf.io/vz2qh>
}
\usage{
consent
}
\description{
A subset of the data from MacDougall (2022) related to the sexual consent behavior of undergraduate students.
}
\details{
This dataset only includes the records for which the number of sexual partners exceeds 0. For the data we obtained from OSF, there were 32 participants who had 0 sexual partners, which exceeds the number reported in the original manuscript, hence some results are slightly different.
}
\keyword{datasets}
