% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drinking}
\alias{drinking}
\title{Drinking and Sweets Dataset}
\format{
A data frame with 150 rows and 13 variables:
\describe{
  \item{ID}{unique participant identification.}
  \item{Gender}{patient self-identified gender.}
  \item{Race}{patient self-identified race.}
  \item{AUDIT}{score on the Alcohol Use Disorder Identification Test (0-40). Higher values indicate more harmful patterns of alcohol intake. Participants had to score at least an 8 to participate in the study, which indicates they were "at-risk drinkers."}
  \item{Height}{height (inches).}
  \item{Weight}{weight (lbs).}
  \item{BMI}{body mass index.}
  \item{Condition}{indicator of treatment group. This is the type of food they were given after receiving an alcohol cue.}
  \item{Craving_Base}{index of alcohol craving on a sliding visual scale (0-100) prior to receiving an alcohol cue. Higher values indicate stronger desire for alcohol.}
  \item{Craving_T1}{index of alcohol craving on a sliding visual scale (0-100) immediately after receiving an alcohol cue. Higher values indicate stronger desire for alcohol.}
  \item{Craving_T2}{index of alcohol craving on a sliding visual scale (0-100) after consuming food following the receipt of an alcohol cue. Higher values indicate stronger desire for alcohol.}
  \item{CravingDiff}{difference in \code{Craving_T2} and \code{Craving_T1}.}
  \item{AlcoholPref}{indicator of preferred alcoholic beverage.}
}
}
\source{
<https://osf.io/4qnvh>
}
\usage{
drinking
}
\description{
Subset of data from a study examining the impact of sweet foods in delaying an alcohol craving among at-risk drinkers.
}
\details{
This dataset combined the two "sweet" groups and the two "bland" groups from the original study and eliminated the control group.
}
\keyword{datasets}
