% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exercise}
\alias{exercise}
\title{Student Exercise Dataset}
\format{
A data frame with 2397 rows and 10 variables:
\describe{
  \item{exercise}{total number of days student exercised during the first two weeks of class.}
  \item{gender}{self-identified gender of student.}
  \item{age}{age of student (years).}
  \item{class}{academic class of student.}
  \item{perceivedSES}{self-perceived socio-economic status.}
  \item{extraversion}{extraversion score from the Big Five Inventory (8-40). Larger values indicate tendency toward extraversion instead of intraversion.}
  \item{agreeableness}{agreeableness score from the Big Five Inventory (9-45). Larger values indicate tendency toward agreeableness instead of antagonism.}
  \item{conscientiousness}{conscientiousness score from the Big Five Inventory (9-45). Larger values indicate tendency toward direction in life instead of lack of direction.}
  \item{neuroticism}{neuroticism score from the Big Five Inventory (8-40). Larger values indicate a tendency toward neuroticism instead of emotional stability.}
  \item{openness}{openness score from the Big Five Inventory (10-50). Larger values indicate an openness instead of closedness towards new experiences.}
}
}
\source{
<https://osf.io/37fgy>
}
\usage{
exercise
}
\description{
Subset of data from a study examining the exercise habits of college students over a semester.
}
\details{
We only consider undergraduate students. The original dataset did not code the BFI domains as totals across the questions, but we have done that here to be consistent with scoring guides.
}
\keyword{datasets}
