% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gratitude}
\alias{gratitude}
\title{Gratitude and Mental Health Dataset}
\format{
A data frame with 123 rows and 9 variables:
\describe{
  \item{Age}{age of participant, in years.}
  \item{Gender}{self-identified gender of participant.}
  \item{Activity}{indicator of physical activity during lockdown protocols.}
  \item{Status}{indicator of socioeconomic status.}
  \item{Wellbeing}{score on Warwick-Edinburgh Mental Well-being Scale (14-70). Higher values indicate a higher level of wellbeing.}
  \item{Gratitude}{score on Gratitude Questionnaire-Six-Item Form (6-42). Higher values indicate higher level of gratitude.}
  \item{Optimism}{score on Life Acceptance Measure (9-45). Higher values indicate higher levels of optimism in the face of tragedy.}
  \item{Support}{score on the Multidimensional Scale of Perceived Social Support (12-84). Higher values indicate higher levels of perceived support.}
  \item{Nature}{score from a few questions asking about nature (3-15).  Higher scores indicate stronger connection with nature.}
}
}
\source{
<https://osf.io/wf5th>
}
\usage{
gratitude
}
\description{
Data from a study examining the insights from the COVID-19 pandemic on strategies for protecting mental well-being.
}
\keyword{datasets}
