% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_formats.R
\name{ma382_word_format}
\alias{ma382_word_format}
\title{Custom Word template.}
\usage{
ma382_word_format(...)
}
\arguments{
\item{...}{additional arguments provided to [rmarkdown::word_document()].}
}
\description{
Loads additional style file for Word for MA223 and MA382.
}
