% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meditation}
\alias{meditation}
\title{Meditation Dataset}
\format{
A data frame with 1281 rows and 12 variables:
\describe{
  \item{respondentID}{unique participant identifier.}
  \item{age}{age of participant (years).}
  \item{gender}{self-identified gender.}
  \item{residence}{content of residence for participant.}
  \item{university}{indicator of whether participant has a university degree.}
  \item{religious}{indicator of whether participant considers themselves to be religious.}
  \item{experience}{years of experience participant has with meditation.}
  \item{frequency}{typical number of times per week participant practices meditation.}
  \item{duration}{typical length of a meditation session for the participant.}
  \item{negthinking}{score from Perseverative Thinking Questionnaire (integers, 0-60). Higher values indicate a higher frequency of repetitive negative thoughts.}
  \item{compassion}{score from Self-Compassion Scale (real number, 1-5). Higher values are indicative of higher levels of self-compassion.}
  \item{mindfulness}{score from Mindsens (real number, 1-5). Higher scores are indicative of higher levels of self-reported mindfulness.}
}
}
\source{
<https://osf.io/gme94>
}
\usage{
meditation
}
\description{
Data from a study examining the impact of meditation on repeated negative thinking.
}
\keyword{datasets}
