% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MilkBacteria}
\alias{MilkBacteria}
\title{Example 4.4 from Montgomery (2008)}
\format{
A data frame with 12 rows and 3 variables:
\describe{
  \item{Day}{indicator of the day on which the corresponding response was collected.}
  \item{Solution}{indicator of the washing solution utilized.}
  \item{CFU}{the number of colony-forming units (CFU's) of bacteria.}
}
}
\source{
Montgomery \emph{Design and Analysis of Experiments} (2008); Exercise 4.4
}
\usage{
MilkBacteria
}
\description{
A dataset containing the bacteria count in several milk containers.
}
\keyword{datasets}
