% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_summaries.R
\name{n}
\alias{n}
\title{Compute sample size.}
\usage{
n(x)
}
\arguments{
\item{x}{any vector.}
}
\description{
This is just an alias for \code{length(x)}.
}
\examples{
summarize_variable(am ~ 1, data = mtcars, n)

}
