% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_summaries.R
\name{nmiss}
\alias{nmiss}
\title{Compute number of missing values.}
\usage{
nmiss(x)
}
\arguments{
\item{x}{any vector.}
}
\description{
This is just an alias for \code{sum(is.na(x))}.
}
\examples{
summarize_variable(am ~ 1, data = mtcars, nmiss)

}
