% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notes}
\alias{notes}
\title{Laptop vs. Longhand Note-taking}
\format{
A data frame with 65 rows and 13 variables:
\describe{
  \item{wordcount}{number of words in the notes taken by the participant.}
  \item{condition}{manner in which participant took notes during lecture.}
  \item{preference}{manner in which participant typically prefers taking notes.}
  \item{class}{indicator of academic class standing.}
  \item{talk}{indicator of subject of the lecture viewed.}
  \item{GPA}{cumulative GPA of participant.}
  \item{SATreading}{score on the reading portion of the SAT (perfect score is 800).}
  \item{SATmath}{score on the math portion of the SAT (perfect score is 800).}
  \item{SATwriting}{score on the writing portion of the SAT (perfect score is 800).}
  \item{SATcombined}{overall SAT score (perfect score is 2400).}
  \item{perception}{participant's perception of whether taking notes with a laptop or longhand is better for students. On a scale of 1-7, with 1 indicating laptop is significantly better and 7 indicating longhand is significantly better.}
  \item{objective}{score on questions assessing recall of facts from the lecture. Potential score varies by lecture topic.}
  \item{conceptual}{score on conceptual open-ended questions. Potential score varies by lecture topic.}
}
}
\source{
<https://osf.io/qrs5y>
}
\usage{
notes
}
\description{
Data from a study comparing note-taking longhand or with a laptop.
}
\keyword{datasets}
