% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_phrases.R
\name{obtain_diagnostics}
\alias{obtain_diagnostics}
\title{Obtain residuals and fitted values for diagnostics.}
\usage{
obtain_diagnostics(mean.model, data)
}
\arguments{
\item{mean.model}{\code{lm} or \code{glm} model fit defining the model from
which residuals and predicted values are to be computed.}

\item{data}{the original data set to which the fitted values and residuals
should be added. Defaults to extracting the data from the model.}
}
\value{
A \code{data.frame} containing one row for each observations, with
two columns added to the original data:
\itemize{
  \item \code{.fitted}: the fitted values for the model.
  \item \code{.resid}: the residuals for the model.}
}
\description{
A wrapper for [generics::augment()] which obtains residuals and
fitted values only for a specified data generating process.
}
\examples{
fit <- lm(mpg ~ hp, data = mtcars)
mtcars.aug <- obtain_diagnostics(fit, data = mtcars)

}
