% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_summaries.R
\name{perc01}
\alias{perc01}
\alias{perc05}
\alias{perc10}
\alias{perc20}
\alias{perc80}
\alias{perc85}
\alias{perc90}
\alias{perc95}
\alias{perc99}
\alias{Q1}
\alias{Q2}
\alias{IQR}
\title{Compute specific percentiles of a numeric variable.}
\usage{
perc01(x, na.rm = FALSE, type = 7)

perc05(x, na.rm = FALSE, type = 7)

perc10(x, na.rm = FALSE, type = 7)

perc20(x, na.rm = FALSE, type = 7)

perc80(x, na.rm = FALSE, type = 7)

perc85(x, na.rm = FALSE, type = 7)

perc90(x, na.rm = FALSE, type = 7)

perc95(x, na.rm = FALSE, type = 7)

perc99(x, na.rm = FALSE, type = 7)

Q1(x, na.rm = FALSE, type = 7)

Q2(x, na.rm = FALSE, type = 7)

IQR(x, na.rm = FALSE, type = 7)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted.  \code{NA} and
\code{NaN} values are not allowed in numeric vectors unless \code{na.rm} is
\code{TRUE}.}

\item{na.rm}{logical; if \code{TRUE}, any \code{NA} or \code{NaN}'s are
removed from \code{x} before the quantiles are computed.}

\item{type}{an integer between 1 and 9 selecting one of the nine quantile
algorithms detailed in \code{\link[stats]{quantile}}.}
}
\description{
These are wrappers for the more common \code{\link[stats]{quantile}} function
which computes the most commonly used percentiles. Functions beginning with
\code{perc} compute the corresponding percentile; functions beginning with
\code{Q} compute the corresponding quartile, and \code{IQR} computes the
interquartile range.
}
\section{Functions}{
\itemize{
\item \code{perc05}: 

\item \code{perc10}: 

\item \code{perc20}: 

\item \code{perc80}: 

\item \code{perc85}: 

\item \code{perc90}: 

\item \code{perc95}: 

\item \code{perc99}: 

\item \code{Q1}: 

\item \code{Q2}: 

\item \code{IQR}: 
}}

\examples{
summarize_variable(mpg ~ 1, data = mtcars, perc01)

}
