% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_summaries.R
\name{percent}
\alias{percent}
\title{Compute percent for a binary variable.}
\usage{
percent(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector or logical vector for which the percent of 1's is
desired.  \code{NA} and \code{NaN} values are not allowed in numeric vectors
unless \code{na.rm} is \code{TRUE}.}

\item{na.rm}{logical; if \code{TRUE}, any \code{NA} or \code{NaN}'s are
removed from \code{x} before the quantiles are computed.}
}
\description{
This is just shorthand for taking \code{100*mean(x)}.
}
\examples{
summarize_variable(am ~ 1, data = mtcars, percent)

}
