% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_null_distribution.R
\name{plot_null_distribution}
\alias{plot_null_distribution}
\title{Plot the null distribution of a test statistic.}
\usage{
plot_null_distribution(statistics, show.pvalue = FALSE)
}
\arguments{
\item{statistics}{result of a call to \code{\link{compare_models}}.}

\item{show.pvalue}{boolean indicating whether the p-value region should
be drawn on the graphic.  If \code{TRUE}, the region is drawn; if
\code{FALSE} (default), the region is not drawn.}
}
\value{
\code{ggplot2} object.
}
\description{
Provides a wrapper for plotting a model for the null distribution of a test
statistic comparing two linear models.
}
\details{
The \code{\link{compare_models}} function returns an attribute that
contains a random sample from the null distribution of the test statistic.
This is extracted and used to construct a graphic summarizing this
distribution.
}
\examples{
fit1 <- lm(mpg ~ hp, data = mtcars)
fit0 <- lm(mpg ~ 1, data = mtcars)

null_distn <- compare_models(fit1, fit0,
  assume.identically.distributed = TRUE,
  assume.normality = FALSE)

plot_null_distribution(null_distn,
  show.pvalue = TRUE)

}
\seealso{
\code{\link{compare_models}}
}
