% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sampling_distribution.R
\name{plot_sampling_distribution}
\alias{plot_sampling_distribution}
\title{Plot the sampling distribution of a statistic.}
\usage{
plot_sampling_distribution(estimates, parameters, conf.int = FALSE)
}
\arguments{
\item{estimates}{result of a call to \code{\link{estimate_parameters}}.}

\item{parameters}{character string indicating for which parameter estimates
the sampling distribution should be plotted. Each will appear in a different
facet of the same graphic. If missing (default), all sampling distributions
are constructed.}

\item{conf.int}{boolean indicating whether the confidence region should be
constructed on the graphic.  If \code{TRUE}, the region is drawn; if
\code{FALSE} (default), the region is not drawn.}
}
\value{
\code{ggplot2} object.
}
\description{
Provides a wrapper for plotting a model for the sampling distribution of a
parameter estimate in a linear model.
}
\details{
The \code{\link{estimate_parameters}} function returns an attribute that
contains a random sample from the sampling distribution of the parameter
estimate.  This is extracted and used to construct a graphic summarizing
this distribution.
}
\examples{
test.df <- data.frame(x = seq(10), y = seq(10) + rnorm(10))
fit <- lm(y ~ x, data = test.df)
samp_distn <- estimate_parameters(fit,
assume.constant.variance = TRUE,
assume.normality = FALSE)
plot_sampling_distribution(samp_distn)

}
\seealso{
\code{\link{estimate_parameters}}
}
