% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RealEstate}
\alias{RealEstate}
\title{Real Estate dataset from Kutner (2005)}
\format{
A data frame with 522 rows and 13 variables:
\describe{
  \item{ID}{Unique identification for each sale.}
  \item{Price}{Sale price of home ($).}
  \item{Area}{Finished area of residence (square feet).}
  \item{BedRooms}{Number of bedrooms in the residence.}
  \item{BathRooms}{Number of bathrooms in the residence.}
  \item{AC}{Indicator of whether home has central air conditioning (1 = Yes, 0 = No).}
  \item{Garage}{Number of cars that garage will hold.}
  \item{Pool}{Indicator of whether home has a swimming pool (1 = Yes, 0 = No).}
  \item{Year}{Year property was originally constructed.}
  \item{Quality}{Indicator of the quality of the construction (1 = High quality, 2 = Medium, 3 = Low).}
  \item{Style}{Indicator of the archiectural style (mapping unknown).}
  \item{Lot}{Lot size (square feet).}
  \item{Highway}{Indicator of whether home is located near highway (1 = Yes, 0 = No).}
}
}
\source{
\url{http://users.stat.ufl.edu/~rrandles/sta4210/Rclassnotes/data/textdatasets/KutnerData/Appendix C Data Sets/APPENC07.txt}
}
\usage{
RealEstate
}
\description{
A dataset containing characteristics on home sales.
}
\keyword{datasets}
