% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rolemodel}
\alias{rolemodel}
\title{Role Models for Women in STEM}
\format{
A data frame with 72 rows and 16 variables:
\describe{
  \item{participant}{unique participant id.}
  \item{condition}{indicator of treatment group. This represents the way in which participants engaged with mentors.}
  \item{race}{indicator of self-identified race.}
  \item{stemintent1}{score from a questionnaire measuring intentions to pursue a STEM career (real number, 1-7) prior to intervention. Higher values indicate stronger intent.}
  \item{stemintent2}{score from a questionnaire measuring intentions to pursue a STEM career (real number, 1-7) after intervention. Higher values indicate stronger intent.}
  \item{stembelong1}{score from a questionnaire measuring sense of belonging within STEM (real number, 1-7) prior to intervention. Higher values indicate stronger sense of belonging.}
  \item{stembelong2}{score from a questionnaire measuring sense of belonging within STEM (real number, 1-7) after intervention. Higher values indicate stronger sense of belonging.}
  \item{stemid1}{score from a questionnaire measuring self-identification with STEM (real number, 1-7) prior to intervention. Higher values indicate stronger sense of identification.}
  \item{stemid2}{score from a questionnaire measuring self-identification with STEM (real number, 1-7) after intervention. Higher values indicate stronger sense of identification.}
  \item{gnat1}{d' score from Go-No-Go Association Task assessing implicit stereotypes of women in STEM prior to intervention. Higher scores indicate higher association of women and STEM.}
  \item{gnat2}{d' score from Go-No-Go Association Task assessing implicit stereotypes of women in STEM after intervention. Higher scores indicate higher association of women and STEM.}
  \item{explicit1}{score from a questionnaire measuring agreement with STEM stereotypes (real number, 1-7) prior to intervention. Higher scores indicate stronger endorsement of stereotypes.}
  \item{explicit2}{score from a questionnaire measuring agreement with STEM stereotypes (real number, 1-7) after intervention. Higher scores indicate stronger endorsement of stereotypes.}
  \item{gpa}{GPA of participant at the end of the semester.}
  \item{stemgpa}{GPA of participant among STEM classes at end of the semester.}
  \item{othergpa}{GPA of participant among non-STEM classes at end of the semester.}
}
}
\source{
<https://osf.io/j7umw>
}
\usage{
rolemodel
}
\description{
Data from a study examining the impact of women in STEM identifying with role models.
}
\details{
Variable naming conventions in the original dataset were not clear. Summary statistics in the original manuscript were used to identify variables, but some output could not be replicated exactly.  In these cases, educated guesses were made based on the naming conventions.
}
\keyword{datasets}
