% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senic}
\alias{senic}
\title{SENIC dataset from Kutner (2005)}
\format{
A data frame with 113 rows and 12 variables:
\describe{
  \item{ID}{Unique hospital identifer.}
  \item{Length of Stay}{Average length of stay of all patients in hospital (days).}
  \item{Age}{Average age of patients (in years).}
  \item{Infection Risk}{estimated probability of acquiring infection in hospital (percent).}
  \item{Routine Culturing Ratio}{Ratio of number of cultures performed to the number of patients without signs or symptoms of hospital-acquired infection, times 100.}
  \item{Routine Xray Ratio}{Ratio of number of X-rays performed to number of patients without signs or symptoms of pneumonia, times 100.}
  \item{Number of Beds}{number of beds in hospital during study period.}
  \item{Medical School}{Indicator of whether the hospital is associated with a medical school (1 = Yes, 2 = No).}
  \item{Region}{Indicator of the geographic region for hospital (1 = NE, 2 = NC, 3 = S, 4 = W).}
  \item{Average Census}{number of patients per day in hospital during study period.}
  \item{Number Nurses}{number of full-time equivalent registered and licensed practical nurses during study period (number of full time plus one half the number of part time).}
  \item{Available Facilities}{Percent of 35 potential facilities and services that are provided by the hospital.}
}
}
\source{
\url{http://users.stat.ufl.edu/~rrandles/sta4210/Rclassnotes/data/textdatasets/KutnerData/Appendix C Data Sets/APPENC01.txt}
}
\usage{
senic
}
\description{
A dataset containing characteristics on hospitals participating in the SENIC
project.
}
\keyword{datasets}
