% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_summaries.R
\name{summarize_relationship}
\alias{summarize_relationship}
\alias{summarise_relationship}
\title{Compute statistics summarizing a relationship between variables.}
\usage{
summarize_relationship(formula, data = NULL, FUN = cor, ...)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}": a symbolic
description identifying the response and the predictor of interest and
optionally a grouping variable.}

\item{data}{an optional data frame containing the variables appearing in
\code{formula}.}

\item{FUN}{a function taking at least two arguments, in which the first
two arguments are the two variables to be related.}

\item{...}{additional parameters to pass to FUN.}
}
\value{
data.frame containing the statistics of interest
}
\description{
Compute statistics, such as a correlation coefficient, summarizing the
relationship between two variables.
}
\details{
The formula is used to specify the response, the predictor variable and
optionally a grouping variable which is specified in the form
\code{Response ~ Predictor + Group}.
}
\examples{
summarize_relationship(mpg ~ hp, data = mtcars, cor)
summarize_relationship(mpg ~ hp, data = mtcars, cor, method = "spearman")

}
