% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_summaries.R
\name{summarize_variable}
\alias{summarize_variable}
\alias{summarise_variable}
\title{Compute summary statistics for a variable.}
\usage{
summarize_variable(data, formula, ..., .args = list())

summarise_variable(data, formula, ..., .args = list())
}
\arguments{
\item{data}{a data frame containing the variables appearing in
\code{formula}.}

\item{formula}{an object of class "[stats::formula()]": a symbolic
description identifying the response and the predictors of interest.}

\item{...}{a comma separated list of functions, each of which computes a
statistic of interest.}

\item{.args}{a named list of additional arguments to be added to all function
calls.}
}
\value{
data.frame containing the values of interest
}
\description{
Compute summary statistics, such as the mean or median, for a variable,
potentially within various groups.
}
\examples{
summarize_variable(mpg ~ am, data = mtcars, mean, median)

}
