% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Therapy}
\alias{Therapy}
\title{Rehabilitation therapy dataset from Kutner (2005)}
\format{
A data frame with 24 rows and 4 variables:
\describe{
  \item{Subject}{unique patient identifer.}
  \item{Prior Activity Level}{indicator of the activity level of the subject prior to surgery.}
  \item{Recovery Time}{time (days) required to complete physical therapy successfully.}
  \item{Age}{age (years, rounded to nearest tenth) of patient.}
}
}
\source{
\url{http://users.stat.ufl.edu/~rrandles/sta4210/Rclassnotes/data/textdatasets/KutnerData/Chapter 22 Data Sets/CH22PR11.txt}
}
\usage{
Therapy
}
\description{
A dataset containing the recovery times of patients undergoing rehabilitation
therapy.
}
\keyword{datasets}
