% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{IntroAnalysis-package}
\alias{IntroAnalysis}
\alias{IntroAnalysis-package}
\title{Functions for aiding model development in introductory statistics courses.}
\description{
Wrappers are provided that allow a single framework for estimating parameters
and comparing nested models (hypothesis testing). This allows us to place the
entire introductory statistics course under the umbrella of linear and
potentially generalized linear models.
}
\details{
The three primary functions you will utilize from \pkg{IntroAnalysis} are
\code{\link{specify_mean_model}}, \code{\link{estimate_parameters}} and
\code{\link{compare_models}}.
}
\author{
\strong{Maintainer}: Reyes Eric \email{reyesem@rose-hulman.edu}

}
