% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sampling_distribution.R
\name{plot_sampling_distribution}
\alias{plot_sampling_distribution}
\title{Plot the sampling distribution of a statistic.}
\usage{
plot_sampling_distribution(
  estimates,
  parameter,
  show.confidence.interval = FALSE
)
}
\arguments{
\item{estimates}{result of a call to \code{\link{estimate_parameters}}.}

\item{parameter}{character string indicating for which parameter the modeled
sampling distribution should be plotted. This should match the output of the
"term" column in \code{\link{estimate_parameters}} output. If missing
(default), the model for the sampling distribution of the first term is
plotted.}

\item{show.confidence.interval}{boolean indicating whether the confidence
region should be constructed on the graphic.  If \code{TRUE}, the region is
drawn; if \code{FALSE} (default), the region is not drawn.}
}
\value{
\code{ggplot2} object.
}
\description{
Provides a wrapper for plotting a model for the sampling distribution of a
parameter estimate in a linear model.
}
\details{
The \code{\link{estimate_parameters}} function returns an attribute that
contains a random sample from the sampling distribution of the parameter
estimate.  This is extracted and used to construct a graphic summarizing
this distribution.
}
\examples{
fit <- lm(mpg ~ hp, data = mtcars)

samp_distn <- estimate_parameters(fit,
  confidence.level = 0.95,
  assume.identically.distributed = TRUE,
  assume.normality = FALSE)

plot_sampling_distribution(samp_distn,
  show.confidence.interval = TRUE)

}
\seealso{
\code{\link{estimate_parameters}}
}
